refactor(game): move game model access to dedicated repository layer
- Introduced GameRepository and GameResultRepository to encapsulate all direct Mongoose interactions for game documents.
- Migrated model imports from services/models/schemas to repositories/models/schemas across all modules.
- Updated services, guards, and utility files to use new repository structure.
- Adjusted imports and DI for GameService, AuthService, and other consumers.

This change improves separation of concerns and paves the way for easier testing and mocking of game-related operations.
