+++ packages/quiz-service/src/game/controllers/game-result.controller.e2e-spec.ts
@@ -23,7 +23,7 @@ import {
 } from '../../../test-utils/utils'
 import { Quiz } from '../../quiz/repositories/models/schemas'
 import { User, UserModel } from '../../user/repositories'
-import { Game, GameResult, TaskType } from '../services/models/schemas'
+import { Game, GameResult, TaskType } from '../repositories/models/schemas'
 
 describe('GameResultController (e2e)', () => {
   let app: INestApplication
