+++ packages/quiz-service/src/game/exceptions/illegal-task-type.exception.ts
@@ -1,6 +1,6 @@
 import { InternalServerErrorException } from '@nestjs/common'
 
-import { TaskType } from '../services/models/schemas'
+import { TaskType } from '../repositories/models/schemas'
 
 export class IllegalTaskTypeException extends InternalServerErrorException {
   constructor(actual: TaskType, expected: TaskType) {
