+++ packages/quiz-service/src/game/repositories/index.ts
@@ -1,0 +1,2 @@
+export * from './game.repository'
+export * from './game-result.repository'
