+++ packages/quiz-service/src/game/repositories/models/schemas/game-result.schema.ts
