+++ packages/quiz-service/src/game/repositories/models/schemas/game.schema.ts
