+++ packages/quiz-service/src/game/repositories/models/schemas/index.ts
