+++ packages/quiz-service/src/game/repositories/models/schemas/participant.schema.ts
