+++ packages/quiz-service/src/game/services/index.ts
@@ -1,9 +1,7 @@
-export * from './game.repository'
 export * from './game.service'
 export * from './game-event.publisher'
 export * from './game-event.subscriber'
 export * from './game-expiry-scheduler.service'
-export * from './game-result.repository'
 export * from './game-result.service'
 export * from './game-task-transition.service'
 export * from './game-task-transition-scheduler'
