+++ packages/quiz-service/src/game/services/utils/game-event.converter.spec.ts
@@ -31,7 +31,11 @@ import {
   MOCK_TYPE_ANSWER_OPTION_VALUE,
   MOCK_TYPE_ANSWER_OPTION_VALUE_ALTERNATIVE,
 } from '../../../../test-utils/data'
-import { Game, GameDocument, LeaderboardTaskItem } from '../models/schemas'
+import {
+  Game,
+  GameDocument,
+  LeaderboardTaskItem,
+} from '../../repositories/models/schemas'
 
 import { buildHostGameEvent } from './game-event.converter'
