+++ packages/quiz-service/src/game/services/utils/game.converter.ts
@@ -3,7 +3,7 @@ import { v4 as uuidv4 } from 'uuid'
 
 import { Quiz } from '../../../quiz/repositories/models/schemas'
 import { User } from '../../../user/repositories'
-import { Game } from '../models/schemas'
+import { Game } from '../../repositories/models/schemas'
 
 import { buildLobbyTask } from './task.converter'
