+++ packages/quiz-service/src/game/services/utils/game.utils.ts
@@ -1,6 +1,6 @@
 import { GameParticipantType } from '@quiz/common'
 
-import { GameDocument } from '../models/schemas'
+import { GameDocument } from '../../repositories/models/schemas'
 
 /**
  * Checks whether a participant is unique in the list of game participants.
