+++ packages/quiz-service/src/game/services/utils/task.converter.spec.ts
@@ -25,7 +25,7 @@ import {
   QuestionTaskTrueFalseAnswer,
   QuestionTaskTypeAnswerAnswer,
   TaskType,
-} from '../models/schemas'
+} from '../../repositories/models/schemas'
 
 import {
   buildQuestionResultTask,
