+++ packages/quiz-service/src/game/services/utils/task.utils.ts
@@ -1,4 +1,4 @@
-import { GameDocument, TaskType } from '../models/schemas'
+import { GameDocument, TaskType } from '../../repositories/models/schemas'
 
 /**
  * Checks if the current task of the game document is a lobby task.
